requirements.txt
requests==2.32.4
sentence-transformers==2.6.1
faiss-cpu==1.8.0 
numpy>=1.24.0
